requirements.txt
requests
python-dotenv
facenet-pytorch
pillow
